Base64 encode secret string_data for comparisons
During creation and patching operations, existing objects are pulled
down and compared to what is about to be applied. If no differences
exist than no change is necessary.

Secrets have a convenience parameter of `string_data` which will take a
string and do a base64 conversion replacing any keys in `data`. When the
existing object is pulled down `string_data` does not exist yet might be
in the object to be patched. This might result in a change that is
unnecessary.

Instead, before doing the comparison, encode anything in `string_data`
to base64 and move it into `data` so that proper comparisons can occur.