Always set apiVersion and kind for resources within a resource list (#298)
* Always set apiVersion and kind for resources within a resource list

When the Kubernetes REST API returns a list of objects, it returns a
resource list object with an apiVersion and kind set to "KindList" where
"Kind" is the Kind of object queried, rather than an simple array of
objects.

The `items` property contains the array of resources returned.  Each
resource in this list does not contain the GVK information, only the
resource list has an apiVersion or kind property set.

This ensures when a resource list is returned, each resource within the
list has it's apiVersion and kind set, to make it similar to other
Kubernetes clients, and easier to work with the returned resources.

* Just do one big serialization at the end of a ResourceList operation

(cherry picked from commit 68cc0e3a73c16928ea9e4b7fc2fff828beab47e3)