Apply should work if resources change in between
In the scenario:
* apply a deployment with replicas set to 1
* scale the deployment to 0 replicas
* apply the deployment with some changes, but replicas set to 1

The deployment should be applied with replicas set to 1.

Fixing that problem had a massive knock on effect to how we manage lists,
but also likely removed a number of more subtle bugs.

We use strategic merge information to better merge lists now,
and increase the number of tests in the test suite to validate that