Take last applied configuration into account when patching
If, as previously, we ignore the last-applied annotation when
comparing desired and actual, we end up with a result where,
when deleting all annotations in a change, no annotations are
deleted(!)

Going from
```
metadata:
  name: my-pod
  annotations:
    x: y
```

to:
```
metadata:
  name: my-pod
```

used to result in a patch that updated the last applied configuration
value only.

Now it correctly sets the old annotations to None too.

Break out apply_patch as a separate method to allow it be tested in
isolation.

Ensure that the last-applied annotation is sorted.

Ensure that the last-applied annotation is loaded consistently (i.e.
without unnecessary conversion to u'value's)